+++ ci/entrypoint.sh
@@ -67,5 +67,8 @@ window.config = {
   glitchtip: {
     dsn: "${SENTRY_DSN}",
   },
+  settings: {
+	waitingRoomDefaultValue: ${WAITING_ROOM_DEFAULT_VALUE}
+  }
 }
 EOF